Fix routes
