+++ academy-ui/src/app/guides/view/[...guideIdAndStep]/page.tsx
@@ -7,7 +7,7 @@ import PageWrapper from '@dodao/web-core/components/core/page/PageWrapper';
 import axios from 'axios';
 
 type GuideViewProps = {
-  params: { guideIdAndStep: string[] };
+  params: Promise<{ guideIdAndStep: string[] }>;
 };
 
 export async function generateMetadata({ params }: GuideViewProps): Promise<Metadata> {
@@ -29,7 +29,7 @@ export async function generateMetadata({ params }: GuideViewProps): Promise<Meta
 }
 
 const GuideView = async ({ params }: GuideViewProps) => {
-  const { guideIdAndStep } = params;
+  const { guideIdAndStep } = await params;
   const space = (await getSpaceServerSide())!;
 
   const response = await axios.get(`${getBaseUrl()}/api/guide/${guideIdAndStep[0]}`);
