+++ academy-ui/src/app/tidbit-collections/view/[tidbitCollectionId]/[tidbitId]/page.tsx
@@ -3,16 +3,17 @@ import { SpaceTypes } from '@/types/space/SpaceDto';
 import { getSpaceServerSide } from '@/utils/space/getSpaceServerSide';
 import React from 'react';
 
-export default async function ByteDetails(props: { params: { tidbitCollectionId: string; tidbitId?: string } }) {
+export default async function ByteDetails({ params }: { params: Promise<{ tidbitCollectionId: string; tidbitId?: string }> }) {
   const space = (await getSpaceServerSide())!;
 
+  const { tidbitCollectionId, tidbitId } = await params;
   return (
     <ViewByteModal
       space={space}
-      byteCollectionId={props.params.tidbitCollectionId}
-      selectedByteId={props.params.tidbitId!}
+      byteCollectionId={tidbitCollectionId}
+      selectedByteId={tidbitId!}
       viewByteModalClosedUrl={space.type === SpaceTypes.TidbitsSite ? '/' : '/tidbit-collections'}
-      afterUpsertByteModalClosedUrl={space.type === SpaceTypes.TidbitsSite ? '/' : `/tidbit-collections/view/${props.params?.tidbitCollectionId}`}
+      afterUpsertByteModalClosedUrl={space.type === SpaceTypes.TidbitsSite ? '/' : `/tidbit-collections/view/${tidbitCollectionId}`}
     />
   );
 }
